backend/requirements.txt
django==2.1.5
djangorestframework==3.9.0
djangorestframework-jwt
python-decouple==3.1
mysqlclient
boto3
django-storages
pillow
requests
django-cors-headers
tqdm
colorama
Unidecode==1.0.23
django-filter
bs4
lxml
tqdm
colorama
